Split mappings.ts
Threshold Subgraph recieves events from several sources (contracts). So
a logical division of events handlers can be each contract.

To achieve this, the generic "mappings.ts" file have been splitted in
"pre-app.ts" and "staking.ts".
